Allow long back swipe to kill app (with new killapp implementation)
We can kang what sysui assistant code does to check if launcher is running,
using a TaskStackChangeListener to always be updated about the running
task id and its component name. This is a solid and reliable way
to easily kill the task and even remove it from recents without the
risk to kill launcher or a different app than the foreground one.

PS: all the code that this commit adds to EdgeBackGestureHandler
can be moved for example in navigation bar view to propagate
running task infos somewhere else. That can be useful for people who
wanna extend the kill app function to non gestural navbar modes
(booooooooooooo those guys, booooooooooooooooooooooooooooo :D )

Change-Id: I491466191c76fa57e31439c222b9455204986ad6